// web/src/api/generated/model/payrollSettingsInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Baba.Api | v1
 * OpenAPI spec version: 1.0.0
 */

export interface PayrollSettingsInput {
  autoPost: boolean;
  nationalEmployeePercent: number;
  nationalEmployerPercent: number;
  foreignEmployeePercent: number;
  foreignEmployerPercent: number;
  /** @nullable */
  insuranceFloor: number | null;
  /** @nullable */
  insuranceCeiling: number | null;
  /** @nullable */
  salaryExpenseAccountId: string | null;
  /** @nullable */
  salariesPayableAccountId: string | null;
  /** @nullable */
  insuranceExpenseAccountId: string | null;
  /** @nullable */
  insurancePayableAccountId: string | null;
  /** @nullable */
  endOfServiceExpenseAccountId: string | null;
  /** @nullable */
  endOfServiceProvisionAccountId: string | null;
}